backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.4
gunicorn==21.2.0
python-dotenv==1.0.0
python-socketio==5.9.0
python-engineio==4.7.1
python-dotenv==1.0.0
Werkzeug==2.3.6
